requirements.txt
#django
Django==3.0
django-filter==2.0.0
django-filters==0.1.0
# Django REST Framework
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
coreapi==2.3.3  # https://github.com/core-api/python-client
django-cors-headers # https://pypi.org/project/django-cors-headers/
drf_yasg
mysqlclient
pyexcel_xlsx
django-crontab
django-money